requirements.txt
flask 
flask-sqlalchemy 
beautifulsoup4 
requests
pydub
streamlit
pandas
altair
matplotlib
librosa
SoundFile
torchaudio
munch
torch
pydub
pyyaml
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
git+https://github.com/resemble-ai/monotonic_align.git